refactor: #27 リクエスト・レスポンススキーマを別ファイル化
